Feature/team service

+++ src/main/java/com/study/studypal/team/service/api/impl/TeamServiceImpl.java
@@ -1,25 +1,26 @@
-package com.study.studypal.team.service.impl;
+package com.study.studypal.team.service.api.impl;
 
 import com.study.studypal.common.dto.ActionResponseDto;
 import com.study.studypal.team.dto.Team.request.CreateTeamRequestDto;
 import com.study.studypal.team.dto.Team.request.UpdateTeamRequestDto;
 import com.study.studypal.team.dto.Team.response.*;
 import com.study.studypal.team.entity.Team;
-import com.study.studypal.team.service.TeamInternalService;
+import com.study.studypal.team.entity.TeamUser;
+import com.study.studypal.team.enums.TeamRole;
+import com.study.studypal.team.service.internal.TeamMembershipInternalService;
 import com.study.studypal.user.entity.User;
 import com.study.studypal.common.exception.BusinessException;
 import com.study.studypal.common.exception.NotFoundException;
 import com.study.studypal.team.repository.TeamRepository;
 import com.study.studypal.common.service.CodeService;
 import com.study.studypal.common.service.FileService;
-import com.study.studypal.team.service.TeamService;
+import com.study.studypal.team.service.api.TeamService;
 import com.study.studypal.common.util.FileUtils;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
-import org.modelmapper.TypeToken;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -34,8 +35,9 @@ import java.util.UUID;
 @Service
 @Transactional
 @RequiredArgsConstructor
-public class TeamServiceImpl implements TeamService, TeamInternalService {
+public class TeamServiceImpl implements TeamService {
     private final TeamRepository teamRepository;
+    private final TeamMembershipInternalService teamMembershipService;
     private final CodeService codeService;
     private final FileService fileService;
     private final ModelMapper modelMapper;
@@ -67,7 +69,8 @@ public class TeamServiceImpl implements TeamService, TeamInternalService {
                         .build();
 
                 teamRepository.save(team);
-
+                teamMembershipService.createMembership(team.getId(), userId, TeamRole.CREATOR);
+                
                 return modelMapper.map(team, TeamResponseDto.class);
             }
             catch (DataIntegrityViolationException e){
@@ -83,7 +86,14 @@ public class TeamServiceImpl implements TeamService, TeamInternalService {
                 ()->new NotFoundException("Team not found.")
         );
 
-        return modelMapper.map(team, TeamOverviewResponseDto.class);
+        TeamOverviewResponseDto overview = modelMapper.map(team, TeamOverviewResponseDto.class);
+
+        TeamUser membership = teamMembershipService.getMemberShip(teamId, userId);
+        if(membership.getRole() == TeamRole.MEMBER) {
+            overview.setTeamCode(null);
+        }
+
+        return overview;
     }
 
     @Override
@@ -107,14 +117,19 @@ public class TeamServiceImpl implements TeamService, TeamInternalService {
     public ListTeamResponseDto getUserJoinedTeams(UUID userId, LocalDateTime cursor, int size) {
         Pageable pageable = PageRequest.of(0, size);
 
-        List<Team> teams = teamRepository.findUserJoinedTeamWithCursor(userId, cursor, pageable);
-        List<TeamSummaryResponseDto> summaries = modelMapper.map(teams, new TypeToken<List<TeamSummaryResponseDto>>() {}.getType());
-
+        List<TeamSummaryResponseDto> teams = teamRepository.findUserJoinedTeamWithCursor(userId, cursor, pageable);
         long total = teamRepository.countUserJoinedTeam(userId);
 
+        LocalDateTime nextCursor = null;
+        if(!teams.isEmpty()) {
+            UUID lastTeamId = teams.get(teams.size() - 1).getId();
+            nextCursor = teamMembershipService.getUserJoinedTeamsListCursor(userId, lastTeamId, teams.size(), size);
+        }
+
         return ListTeamResponseDto.builder()
-                .teams(summaries)
+                .teams(teams)
                 .total(total)
+                .nextCursor(nextCursor)
                 .build();
     }
 
@@ -123,14 +138,19 @@ public class TeamServiceImpl implements TeamService, TeamInternalService {
         String handledKeyword = keyword.toLowerCase().trim();
         Pageable pageable = PageRequest.of(0, size);
 
-        List<Team> teams = teamRepository.searchUserJoinedTeamByNameWithCursor(userId, handledKeyword, cursor, pageable);
-        List<TeamSummaryResponseDto> summaries = modelMapper.map(teams, new TypeToken<List<TeamSummaryResponseDto>>() {}.getType());
-
+        List<TeamSummaryResponseDto> teams = teamRepository.searchUserJoinedTeamByNameWithCursor(userId, handledKeyword, cursor, pageable);
         long total = teamRepository.countUserJoinedTeamByName(userId, handledKeyword);
 
+        LocalDateTime nextCursor = null;
+        if(!teams.isEmpty()) {
+            UUID lastTeamId = teams.get(teams.size() - 1).getId();
+            nextCursor = teamMembershipService.getUserJoinedTeamsListCursor(userId, lastTeamId, teams.size(), size);
+        }
+
         return ListTeamResponseDto.builder()
-                .teams(summaries)
+                .teams(teams)
                 .total(total)
+                .nextCursor(nextCursor)
                 .build();
     }
 
@@ -139,6 +159,8 @@ public class TeamServiceImpl implements TeamService, TeamInternalService {
         Team team = teamRepository.findById(teamId).orElseThrow(
                 () -> new NotFoundException("Team not found.")
         );
+
+        teamMembershipService.validateUpdateTeamPermission(userId, teamId);
 
         if(request.getName() != null) {
             if(request.getName().isEmpty()) {
@@ -165,6 +187,8 @@ public class TeamServiceImpl implements TeamService, TeamInternalService {
                 () -> new NotFoundException("Team not found.")
         );
 
+        teamMembershipService.validateUpdateTeamPermission(userId, teamId);
+
         String teamCode = codeService.generateTeamCode();
         while(teamRepository.existsByTeamCode(teamCode)){
             teamCode = codeService.generateTeamCode();
@@ -185,6 +209,8 @@ public class TeamServiceImpl implements TeamService, TeamInternalService {
                 ()-> new NotFoundException("Team not found.")
         );
 
+        teamMembershipService.validateUpdateTeamPermission(userId, teamId);
+
         if(team.getAvatarUrl() != null) {
             fileService.deleteFile(team.getId().toString(), "image");
         }
@@ -203,6 +229,8 @@ public class TeamServiceImpl implements TeamService, TeamInternalService {
             throw new BusinessException("Team's avatar must be an image.");
         }
 
+        teamMembershipService.validateUpdateTeamPermission(userId, teamId);
+
         try {
             String avatarUrl = fileService.uploadFile(AVATAR_FOLDER, teamId.toString(), file.getBytes()).getUrl();
             Team team = teamRepository.findById(teamId).orElseThrow(
@@ -219,46 +247,6 @@ public class TeamServiceImpl implements TeamService, TeamInternalService {
 
         } catch (IOException e) {
             throw new BusinessException("Reading file failed.");
-        }
-    }
-
-
-
-    @Override
-    public UUID getTeamIdByTeamCode(String teamCode) {
-        Team team = teamRepository.findByTeamCode(teamCode);
-
-        if(team == null) {
-            throw new NotFoundException("Team code is incorrect.");
-        }
-
-        return team.getId();
-    }
-
-    @Override
-    public void increaseMember(UUID teamId) {
-        Team team = teamRepository.findById(teamId).orElseThrow(
-                () -> new NotFoundException("Team not found.")
-        );
-
-        team.setTotalMembers(team.getTotalMembers() + 1);
-        teamRepository.save(team);
-    }
-
-    @Override
-    @Transactional
-    public void decreaseMember(UUID teamId) {
-        Team team = teamRepository.findById(teamId).orElseThrow(
-                () -> new NotFoundException("Team not found.")
-        );
-
-        team.setTotalMembers(team.getTotalMembers() - 1);
-
-        if(team.getTotalMembers() == 0) {
-            teamRepository.delete(team);
-        }
-        else {
-            teamRepository.save(team);
         }
     }
 }
